pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rekordbox-bulk-edit"
version = "0.1.0"
description = "Tools for bulk editing RekordBox database records and associated track files"
authors = [{ name = "James Viall", email = "jamesviall@pm.me" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
keywords = ["rekordbox", "dj", "music", "database", "bulk", "edit", "convert"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Multimedia :: Sound/Audio",
]
dependencies = ["pyrekordbox>=0.1.0", "ffmpeg-python>=0.2.0", "click>=8.0.0"]

[project.optional-dependencies]
dev = ["pytest>=6.0", "black", "flake8", "mypy"]

[project.urls]
Homepage = "https://github.com/jviall/rekordbox-bulk-edit"
Repository = "https://github.com/jviall/rekordbox-bulk-edit"
Issues = "https://github.com/jviall/rekordbox-bulk-edit/issues"

[project.scripts]
rekordbox-bulk-edit = "rekordbox_bulk_edit.cli:main"
